backend/requirements.txt
fastapi
uvicorn
pymongo
langchain
chromadb
openai
python-dotenv
sqlmodel
passlib[bcrypt]
python-jose